Added a leaderboard and fixed some previous issues
Fixed an issue when you run out of targets, and you have to wait till the end of your dedicated time to restart. (Also added a special message on the lose panel, to tell the player, what's wrong)
Added a leaderboard, that shows 10 best player's tries is a cute little table. This data is serealized to json and saved in PlayerPrefs.